Refactor the wayland helpers into WaylandApp and wayland_shm
