+++ examples/example-server-lib/wayland_shm.h
@@ -1,0 +1,36 @@
+/*
+ * Copyright © 2018 Canonical Ltd.
+ *
+ * This program is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 or 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * Authored by: Alan Griffiths <alan@octopull.co.uk>
+ */
+
+#ifndef MIRAL_WAYLAND_SHM_H
+#define MIRAL_WAYLAND_SHM_H
+
+#include <wayland-client.h>
+#include <memory>
+#include <vector>
+#include <functional>
+#include <unordered_map>
+
+template<typename Type>
+auto make_scoped(Type* owned, void(*deleter)(Type*)) -> std::unique_ptr<Type, void(*)(Type*)>
+{
+    return {owned, deleter};
+}
+
+wl_shm_pool* make_shm_pool(struct wl_shm* shm, int size, void **data);
+
+#endif // MIRAL_WAYLAND_SHM_H
